+++ src/types/block.ts
@@ -1,0 +1,6 @@
+// TODO: notion apiから取ってこれる全てのblock要素をここで定義したい。
+// 参考：https://developers.notion.com/reference/block
+
+export type Block = any;
+
+export type RichText = any;
